+++ CHANGELOG.md
@@ -6,6 +6,10 @@
   `kubectl` binary, which might not be present for OpenShift users
   (who have `oc` instead of `kubectl`).
 
+- Feature: `skipLogin` can be used in the config.yml to tell the cli not to connect to cloud when using an air-gapped environment.
+
+- Change: `TELEPRESENCE_AGENT_IMAGE` and `TELEPRESENCE_REGISTRY` are now only configurable via config.yml.
+
 - Bugfix: Fixed and improved several error messages, to hopefully be
   more helpful.
 - Bugfix: Fixed a DNS problem on macOS causing slow DNS lookups when connecting to a local cluster.